disable cache on recovery.

+++ assets/js/services/cacheService.js
@@ -1,7 +1,7 @@
 define(["whispeerHelper", "dexie", "bluebird", "services/serviceModule", "services/errorService"], function (h, Dexie, Promise, serviceModule) {
 	"use strict";
 
-	var db, errorService;
+	var db, errorService, cacheDisabled = false;
 
 	function Cache(name, options) {
 		this._name = name;
@@ -10,10 +10,18 @@ define(["whispeerHelper", "dexie", "bluebird", "services/serviceModule", "servic
 	}
 
 	Cache.prototype.entries = function () {
+		if (cacheDisabled) {
+			return Promise.resolve();
+		}
+
 		return db.cache.where("type").equals(this._name);
 	};
 
 	Cache.prototype.entryCount = function () {
+		if (cacheDisabled) {
+			return Promise.reject();
+		}
+
 		return Promise.resolve(db.cache.where("type").equals(this._name).count());
 	};
 
@@ -27,6 +35,10 @@ define(["whispeerHelper", "dexie", "bluebird", "services/serviceModule", "servic
 	};
 
 	Cache.prototype.store = function (id, data, blob) {
+		if (cacheDisabled) {
+			return Promise.resolve();
+		}
+
 		if (blob && blob.size > 1*1024*1024) {
 			return Promise.resolve();
 		}
@@ -60,6 +72,10 @@ define(["whispeerHelper", "dexie", "bluebird", "services/serviceModule", "servic
 	};
 
 	Cache.prototype.get = function (id) {
+		if (cacheDisabled) {
+			return Promise.reject();
+		}
+
 		var cacheResult = db.cache.where("id").equals(this._name + "/" + id);
 
 		db.cache.where("id").equals(this._name + "/" + id).modify({ used: new Date().getTime() });
@@ -81,6 +97,10 @@ define(["whispeerHelper", "dexie", "bluebird", "services/serviceModule", "servic
 
 	/** get all cache entries as a dexie collection. */
 	Cache.prototype.all = function () {
+		if (cacheDisabled) {
+			return Promise.resolve();
+		}
+
 		return db.cache.where("id").startsWith(this._name + "/");
 	};
 
@@ -88,10 +108,18 @@ define(["whispeerHelper", "dexie", "bluebird", "services/serviceModule", "servic
 	* id: id of the entry
 	*/
 	Cache.prototype.delete = function (id) {
+		if (cacheDisabled) {
+			return Promise.resolve();
+		}
+
 		return db.cache.where("id").equals(this._name + "/" + id).delete();
 	};
 
 	Cache.prototype.cleanUp = function () {
+		if (cacheDisabled) {
+			return Promise.resolve();
+		}
+
 		if (this._options.maxEntries === -1) {
 			return;
 		}
@@ -106,22 +134,12 @@ define(["whispeerHelper", "dexie", "bluebird", "services/serviceModule", "servic
 		}));
 	};
 
-	function NoCache() {}
-
-	NoCache.prototype.entryCount = function () {
-		return Promise.reject();
+	Cache.disable = function () {
+		cacheDisabled = true;
 	};
 
-	NoCache.prototype.store = function () {
-		return Promise.resolve();
-	};
-
-	NoCache.prototype.get = function () {
-		return Promise.reject();
-	};
-
-	NoCache.prototype.cleanUp = function () {
-		return Promise.resolve();
+	Cache.enable = function () {
+		cacheDisabled = false;
 	};
 
 	try {
@@ -144,11 +162,11 @@ define(["whispeerHelper", "dexie", "bluebird", "services/serviceModule", "servic
 	function service (_errorService) {
 		errorService = _errorService;
 
-		if (!db) {
-			return NoCache;
+		if (db) {
+			db.on("blocked", errorService.logError);
+		} else {
+			Cache.disable();
 		}
-
-		db.on("blocked", errorService.logError);
 
 		return Cache;
 	}
